refactor(app): status-bar refactor
